+++ src/api/languages.rs
@@ -2,9 +2,6 @@
 
 use serde::{Deserialize, Serialize};
 
-/// LanguageTool GET languages request.
-pub struct Request;
-
 #[derive(Clone, PartialEq, Eq, Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 #[non_exhaustive]