+++ src/api/server.rs
@@ -1,13 +1,11 @@
 //! Structure to communicate with some `LanguageTool` server through the API.
 
 use crate::{
-    check::{CheckRequest, CheckResponse, CheckResponseWithContext},
-    error::{Error, Result},
-    languages::LanguagesResponse,
-    words::{
-        WordsAddRequest, WordsAddResponse, WordsDeleteRequest, WordsDeleteResponse, WordsRequest,
-        WordsResponse,
+    api::{
+        check::{self, Request, Response},
+        languages, words,
     },
+    error::{Error, Result},
 };
 #[cfg(feature = "cli")]
 use clap::Args;
@@ -25,7 +23,7 @@ use std::{io, path::PathBuf, time::Instant};
 /// # Examples
 ///
 /// ```
-/// # use languagetool_rust::server::parse_port;
+/// # use languagetool_rust::api::server::parse_port;
 /// assert!(parse_port("8081").is_ok());
 ///
 /// assert!(parse_port("").is_ok()); // No port specified, which is accepted
@@ -368,7 +366,7 @@ impl ServerClient {
     }
 
     /// Send a check request to the server and await for the response.
-    pub async fn check(&self, request: &CheckRequest) -> Result<CheckResponse> {
+    pub async fn check(&self, request: &Request) -> Result<Response> {
         match self
             .client
             .post(format!("{0}/check", self.api))
@@ -379,7 +377,7 @@ impl ServerClient {
             Ok(resp) => {
                 match resp.error_for_status_ref() {
                     Ok(_) => {
-                        resp.json::<CheckResponse>()
+                        resp.json::<Response>()
                             .await
                             .map_err(Error::ResponseDecode)
                             .map(|mut resp| {
@@ -410,10 +408,7 @@ impl ServerClient {
     ///
     /// If any of the requests has `self.text` field which is none.
     #[cfg(feature = "multithreaded")]
-    pub async fn check_multiple_and_join(
-        &self,
-        requests: Vec<CheckRequest>,
-    ) -> Result<CheckResponse> {
+    pub async fn check_multiple_and_join(&self, requests: Vec<Request>) -> Result<Response> {
         let mut tasks = Vec::with_capacity(requests.len());
 
         for request in requests.into_iter() {
@@ -423,20 +418,22 @@ impl ServerClient {
                 let text = request.text.ok_or(Error::InvalidRequest(
                     "missing text field; cannot join requests with data annotations".to_string(),
                 ))?;
-                Result::<(String, CheckResponse)>::Ok((text, response))
+                Result::<(String, Response)>::Ok((text, response))
             }));
         }
 
-        let mut response_with_context: Option<CheckResponseWithContext> = None;
+        let mut response_with_context: Option<check::ResponseWithContext> = None;
 
         for task in tasks {
             let (text, response) = task.await.unwrap()?;
             match response_with_context {
                 Some(resp) => {
                     response_with_context =
-                        Some(resp.append(CheckResponseWithContext::new(text, response)))
+                        Some(resp.append(check::ResponseWithContext::new(text, response)))
                 },
-                None => response_with_context = Some(CheckResponseWithContext::new(text, response)),
+                None => {
+                    response_with_context = Some(check::ResponseWithContext::new(text, response))
+                },
             }
         }
 
@@ -448,7 +445,7 @@ impl ServerClient {
     #[cfg(feature = "annotate")]
     pub async fn annotate_check(
         &self,
-        request: &CheckRequest,
+        request: &Request,
         origin: Option<&str>,
         color: bool,
     ) -> Result<String> {
@@ -459,7 +456,7 @@ impl ServerClient {
     }
 
     /// Send a languages request to the server and await for the response.
-    pub async fn languages(&self) -> Result<LanguagesResponse> {
+    pub async fn languages(&self) -> Result<languages::Response> {
         match self
             .client
             .get(format!("{}/languages", self.api))
@@ -469,7 +466,7 @@ impl ServerClient {
             Ok(resp) => {
                 match resp.error_for_status_ref() {
                     Ok(_) => {
-                        resp.json::<LanguagesResponse>()
+                        resp.json::<languages::Response>()
                             .await
                             .map_err(Error::ResponseDecode)
                     },
@@ -481,7 +478,7 @@ impl ServerClient {
     }
 
     /// Send a words request to the server and await for the response.
-    pub async fn words(&self, request: &WordsRequest) -> Result<WordsResponse> {
+    pub async fn words(&self, request: &words::Request) -> Result<words::Response> {
         match self
             .client
             .get(format!("{}/words", self.api))
@@ -492,7 +489,7 @@ impl ServerClient {
             Ok(resp) => {
                 match resp.error_for_status_ref() {
                     Ok(_) => {
-                        resp.json::<WordsResponse>()
+                        resp.json::<words::Response>()
                             .await
                             .map_err(Error::ResponseDecode)
                     },
@@ -504,7 +501,7 @@ impl ServerClient {
     }
 
     /// Send a words/add request to the server and await for the response.
-    pub async fn words_add(&self, request: &WordsAddRequest) -> Result<WordsAddResponse> {
+    pub async fn words_add(&self, request: &words::add::Request) -> Result<words::add::Response> {
         match self
             .client
             .post(format!("{}/words/add", self.api))
@@ -515,7 +512,7 @@ impl ServerClient {
             Ok(resp) => {
                 match resp.error_for_status_ref() {
                     Ok(_) => {
-                        resp.json::<WordsAddResponse>()
+                        resp.json::<words::add::Response>()
                             .await
                             .map_err(Error::ResponseDecode)
                     },
@@ -527,7 +524,10 @@ impl ServerClient {
     }
 
     /// Send a words/delete request to the server and await for the response.
-    pub async fn words_delete(&self, request: &WordsDeleteRequest) -> Result<WordsDeleteResponse> {
+    pub async fn words_delete(
+        &self,
+        request: &words::delete::Request,
+    ) -> Result<words::delete::Response> {
         match self
             .client
             .post(format!("{}/words/delete", self.api))
@@ -538,7 +538,7 @@ impl ServerClient {
             Ok(resp) => {
                 match resp.error_for_status_ref() {
                     Ok(_) => {
-                        resp.json::<WordsDeleteResponse>()
+                        resp.json::<words::delete::Response>()
                             .await
                             .map_err(Error::ResponseDecode)
                     },
@@ -583,7 +583,8 @@ impl ServerClient {
 
 #[cfg(test)]
 mod tests {
-    use crate::{check::CheckRequest, ServerClient};
+    use super::ServerClient;
+    use crate::api::check::Request;
 
     #[tokio::test]
     async fn test_server_ping() {
@@ -594,14 +595,14 @@ mod tests {
     #[tokio::test]
     async fn test_server_check_text() {
         let client = ServerClient::from_env_or_default();
-        let req = CheckRequest::default().with_text("je suis une poupee".to_string());
+        let req = Request::default().with_text("je suis une poupee".to_string());
         assert!(client.check(&req).await.is_ok());
     }
 
     #[tokio::test]
     async fn test_server_check_data() {
         let client = ServerClient::from_env_or_default();
-        let req = CheckRequest::default()
+        let req = Request::default()
             .with_data_str("{\"annotation\":[{\"text\": \"je suis une poupee\"}]}")
             .unwrap();
         assert!(client.check(&req).await.is_ok());